feat: Updated logic for processing component length and quantity in the interface and PDF generation
- Changed data type for component length from float to int to improve accuracy.
- Updated default values ​​for length in the interface and PDF generation.
- Simplified logic for processing length and quantity in the functions for updating and filling fields.

+++ new_project/form_com_offer/utils/pdf_generator.py
@@ -355,7 +355,7 @@ async def generate_commercial_offer_pdf(
                 price = float(comp.get('price', 0))
                 unit = comp.get('unit', 'шт.')
                 if unit == 'м.':
-                    qty_or_length = float(comp.get('length', 0.0))
+                    qty_or_length = int(comp.get('length', 0))
                 else:
                     qty_or_length = int(comp.get('qty', 0))
